requirements.txt
awesome-slugify==1.6.5
certifi==2017.11.5
chardet==3.0.4
Django==1.11.3
django-admin-jqueryui112==1.12.1b5
django-braces==1.12.0
django-cart==1.0.4
django-changuito==1.2
django-cleanup==2.1.0
django-content-gallery==1.0.0b1
django-debug-toolbar==1.9.1
django-paypal==0.4.1
django-widget-tweaks==1.4.1
docopt==0.6.2
easy-thumbnails==2.5
idna==2.6
image==1.5.16
libmagic==1.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pipreqs==0.4.9
python-magic==0.4.13
pytz==2017.2
regex==2017.11.9
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
Unidecode==0.4.21
urllib3==1.22
wincertstore==0.2
yarg==0.1.9
